requirements.txt
pandas
numpy
matplotlib
seaborn
biopython
tqdm
requests
openpyxl
pybiomart
